Logic issue while compiling SeDuMi on Matlab version > 7.3 (`ddot_` error or `typedef mwSize` error)
The `sedumi_install` script contains a logic error in the if conditions for setting up compilation flags. This may bring to a failing compilation on Matlab version with `VERSION(1) > 7` and `VERSION(2) < 3` (i.e. Matlab 2017a in my case).

The compilation flags selected make the compiler to raise cryptic error messages about linking (`ddot_` linking error, that is not so easy to understand only if you don't know blas and lapack libraries), or inform the user about a re-definition of some Matlab size type (in particular it brings in the custom definitions of `mwIndex`, `mwSize` and `mwSignedInt` instead of requiring the `largeArrayDim`).

I strongly suggest to fix the `sedumi_install` script by using the Matlab function `varLessThan` (introduced in version 7.3) in place of a manual comparison, but to keep it compatible with version prior to 7.3, the logic condition has been updated using the `VERSION` variable.